+++ Ui/Component/Listing/Columns/SnakeToTitleCaseRenderer.php
@@ -18,7 +18,7 @@ class SnakeToTitleCaseRenderer extends Column
     /**
      * @inheritDoc
      */
-    public function prepareDataSource(array $dataSource)
+    public function prepareDataSource(array $dataSource): array
     {
         $componentIndex = $this->getData('name');
         foreach ($dataSource['data']['items'] ?? [] as $index => $item) {